examples/260-aws-lambda-python-transcription/requirements.txt
deepgram-sdk==6.1.1
boto3>=1.34.0
